organize cyclepoints into its own folder
