+++ bycycle/cyclepoints/__init__.py
@@ -1,0 +1,4 @@
+"""Cyclepoint functionality."""
+
+from .extrema import find_extrema, find_zerox
+from .phase import extrema_interpolated_phase
